refactor: 테스트 패키지 폴더 구조 정리
