+++ src/test/java/com/odiga/fiesta/FiestaApplicationTests.java
@@ -5,6 +5,8 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.context.annotation.Import;
 import org.springframework.test.context.ActiveProfiles;
 
+import com.odiga.fiesta.config.S3MockConfig;
+
 @ActiveProfiles("test")
 @SpringBootTest
 @Import(S3MockConfig.class)
