+++ src/test/java/com/odiga/fiesta/RepositoryTestSupport.java
@@ -5,6 +5,8 @@ import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import org.springframework.context.annotation.Import;
 import org.springframework.test.context.ActiveProfiles;
 
+import com.odiga.fiesta.config.QuerydslTestConfig;
+
 @DataJpaTest
 @Import({QuerydslTestConfig.class})
 @ActiveProfiles("test")
